change document structure

+++ backend/controllers/textbookRequestController.js
@@ -4,9 +4,9 @@ const Textbook = require("../models/Textbook");
 // Create a new textbook request
 exports.createTextbookRequest = async (req, res) => {
   const { textbooks, userID } = req.body;
+  const year = new Date().getFullYear(); // Set current year
 
   try {
-    // Check if all textbooks exist
     const textbooksExist = await Promise.all(
       textbooks.map(async (item) => {
         const textbook = await Textbook.findById(item.textbook);
@@ -14,7 +14,6 @@ exports.createTextbookRequest = async (req, res) => {
       })
     );
 
-    // If any textbook does not exist, return error
     if (textbooksExist.some((exist) => !exist)) {
       return res
         .status(400)
@@ -24,6 +23,8 @@ exports.createTextbookRequest = async (req, res) => {
     const newRequest = new TextbookRequest({
       textbooks,
       userID,
+      year, // Add year to the request
+      submission: false, // Set submission to false by default
     });
 
     await newRequest.save();
@@ -33,10 +34,13 @@ exports.createTextbookRequest = async (req, res) => {
   }
 };
 
-// Get all textbook requests
+// Get all textbook requests, optionally filtered by year
 exports.getAllTextbookRequests = async (req, res) => {
+  const year = req.query.year ? parseInt(req.query.year) : null;
+
   try {
-    const requests = await TextbookRequest.find()
+    const query = year ? { year } : {};
+    const requests = await TextbookRequest.find(query)
       .populate({
         path: "textbooks.textbook",
         select: "title grade subject language category level",
@@ -49,6 +53,32 @@ exports.getAllTextbookRequests = async (req, res) => {
           select: "name", // Select the name of the organization
         },
       });
+
+    res.status(200).json(requests);
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
+};
+
+// Get all textbook requests for a specific year
+exports.getAllTextbookRequestsByYear = async (req, res) => {
+  const year = parseInt(req.params.year);
+
+  try {
+    const requests = await TextbookRequest.find({ year })
+      .populate({
+        path: "textbooks.textbook",
+        select: "title grade subject language category level",
+      })
+      .populate({
+        path: "userID",
+        select: "full_name organization",
+        populate: {
+          path: "organization",
+          select: "name",
+        },
+      });
+
     res.status(200).json(requests);
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -77,7 +107,7 @@ exports.getTextbookRequestById = async (req, res) => {
         populate: {
           path: "organization", // Populate the organization field
           select: "name",
-         }, // Select full_name of the user who evaluated
+        }, // Select full_name of the user who evaluated
       });
     if (!request) {
       return res.status(404).json({ message: "Textbook request not found" });
@@ -91,14 +121,20 @@ exports.getTextbookRequestById = async (req, res) => {
 // Update a textbook request
 exports.updateTextbookRequest = async (req, res) => {
   try {
-    const request = await TextbookRequest.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      { new: true }
-    );
+    const request = await TextbookRequest.findById(req.params.id);
     if (!request) {
       return res.status(404).json({ message: "Textbook request not found" });
     }
+
+    if (request.submission) {
+      return res
+        .status(400)
+        .json({ message: "Cannot update a submitted request" });
+    }
+
+    Object.assign(request, req.body); // Update the request with new data
+    await request.save();
+
     res.status(200).json(request);
   } catch (err) {
     res.status(400).json({ message: err.message });
@@ -108,10 +144,18 @@ exports.updateTextbookRequest = async (req, res) => {
 // Delete a textbook request
 exports.deleteTextbookRequest = async (req, res) => {
   try {
-    const request = await TextbookRequest.findByIdAndDelete(req.params.id);
+    const request = await TextbookRequest.findById(req.params.id);
     if (!request) {
       return res.status(404).json({ message: "Textbook request not found" });
     }
+
+    if (request.submission) {
+      return res
+        .status(400)
+        .json({ message: "Cannot delete a submitted request" });
+    }
+
+    await TextbookRequest.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: "Textbook request deleted successfully" });
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -130,9 +174,11 @@ exports.updateTextbookRequestEvaluation = async (req, res) => {
 
     // Update the evaluation fields
     request.evaluation.status = status || request.evaluation.status;
-    request.evaluation.evaluatedBy = evaluatedBy || request.evaluation.evaluatedBy;
+    request.evaluation.evaluatedBy =
+      evaluatedBy || request.evaluation.evaluatedBy;
     request.evaluation.comment = comment || request.evaluation.comment;
-    request.evaluation.evaluationDate = evaluationDate || request.evaluation.evaluationDate;
+    request.evaluation.evaluationDate =
+      evaluationDate || request.evaluation.evaluationDate;
 
     await request.save();
 
@@ -141,7 +187,6 @@ exports.updateTextbookRequestEvaluation = async (req, res) => {
     res.status(400).json({ message: err.message });
   }
 };
-
 
 // Update a single textbook within a textbook request
 exports.updateSingleTextbook = async (req, res) => {
@@ -154,9 +199,19 @@ exports.updateSingleTextbook = async (req, res) => {
       return res.status(404).json({ message: "Textbook request not found" });
     }
 
-    const textbook = request.textbooks.find((t) => t.textbook.toString() === textbookId);
+    if (request.submission) {
+      return res
+        .status(400)
+        .json({ message: "Cannot update a textbook in a submitted request" });
+    }
+
+    const textbook = request.textbooks.find(
+      (t) => t.textbook.toString() === textbookId
+    );
     if (!textbook) {
-      return res.status(404).json({ message: "Textbook not found in the request" });
+      return res
+        .status(404)
+        .json({ message: "Textbook not found in the request" });
     }
 
     textbook.quantity = quantity;
@@ -178,6 +233,12 @@ exports.deleteSingleTextbook = async (req, res) => {
       return res.status(404).json({ message: "Textbook request not found" });
     }
 
+    if (request.submission) {
+      return res
+        .status(400)
+        .json({ message: "Cannot delete a textbook from a submitted request" });
+    }
+
     request.textbooks = request.textbooks.filter(
       (t) => t.textbook.toString() !== textbookId
     );
@@ -188,3 +249,47 @@ exports.deleteSingleTextbook = async (req, res) => {
     res.status(400).json({ message: err.message });
   }
 };
+
+// Submit a textbook request
+exports.submitTextbookRequest = async (req, res) => {
+  try {
+    const request = await TextbookRequest.findById(req.params.id);
+    if (!request) {
+      return res.status(404).json({ message: "Textbook request not found" });
+    }
+
+    request.submission = true; // Set submission to true
+    await request.save();
+
+    res.status(200).json(request);
+  } catch (err) {
+    res.status(400).json({ message: err.message });
+  }
+};
+
+
+exports.addManyTextbookRequests = async (req, res) => {
+  const { textbookRequests } = req.body; // Assuming req.body.textbookRequests is an array of textbook request objects
+
+  try {
+    // Use insertMany to insert multiple textbook requests at once
+    const insertedRequests = await TextbookRequest.insertMany(textbookRequests);
+
+    res.status(201).json({
+      success: true,
+      message: "Textbook requests added successfully",
+      textbookRequests: insertedRequests,
+    });
+  } catch (err) {
+    console.error("Error adding textbook requests:", err);
+    res.status(500).json({
+      success: false,
+      message: "Failed to add textbook requests",
+      error: err.message,
+    });
+  }
+};
+
+
+
+
